Extracted VueX store for log messages and stats into separate modules
